new file for checking the commands
