+++ minishell.h
@@ -119,5 +119,6 @@ char * get_the_oldpwd(t_content *cont, int i);
 void handle_execve_error_message(int error_number, t_content *cont, int i);
 int cmd_has_path(t_content *cont, int i);
 void check_for_path(t_content *cont, int i);
+int cmd_is_program(t_content *cont, int i);
 
 #endif
